requirements.txt
langchain
langchain-openai
langchain-community
langsmith
neo4j
python-dotenv 